ROX-26287: Missing deployment info in violations deployment tab
