+++ ui/apps/platform/src/Containers/Violations/Details/Deployment/ContainerConfiguration.tsx
@@ -1,92 +1,38 @@
 import React, { ReactElement } from 'react';
-import { DescriptionList, Flex, FlexItem } from '@patternfly/react-core';
+import { Card, CardBody, CardTitle, Title } from '@patternfly/react-core';
 
-import DescriptionListItem from 'Components/DescriptionListItem';
-import { Container } from 'types/deployment.proto';
+import { Deployment } from 'types/deployment.proto';
+import ContainerConfigurationDescriptionList from './ContainerConfigurationDescriptionList';
 
-import ContainerImage from './ContainerImage';
-import ContainerResourcesDescriptionList from './ContainerResourcesDescriptionList';
-import ContainerSecretDescriptionList from './ContainerSecretDescriptionList';
-import ContainerVolumeDescriptionList from './ContainerVolumeDescriptionList';
-
-function MultilineDescription({ descArr }) {
-    return (
-        <Flex direction={{ default: 'column' }}>
-            {descArr.map((desc, idx) => (
-                // eslint-disable-next-line react/no-array-index-key
-                <div key={idx}>
-                    <FlexItem>{desc}</FlexItem>
-                </div>
-            ))}
-        </Flex>
-    );
-}
-
-type ContainerConfigurationProps = {
-    container: Container;
+export type ContainerConfigurationProps = {
+    deployment: Deployment | null;
 };
 
-function ContainerConfiguration({ container }: ContainerConfigurationProps): ReactElement {
-    const { resources, volumes, secrets, config, image } = container;
-    const { command, args } = config || {};
+function ContainerConfiguration({ deployment }: ContainerConfigurationProps): ReactElement {
+    let content: JSX.Element[] | string = 'None';
+
+    if (deployment === null) {
+        content =
+            "Container configurations are unavailable because the alert's deployment no longer exists.";
+    } else {
+        if (Array.isArray(deployment?.containers) && deployment.containers.length !== 0) {
+            content = deployment.containers.map((container, i) => (
+                <React.Fragment key={container.id}>
+                    <Title headingLevel="h4" className="pf-v5-u-mb-md">{`containers[${i}]`}</Title>
+                    <ContainerConfigurationDescriptionList
+                        key={container.id}
+                        container={container}
+                    />
+                </React.Fragment>
+            ));
+        }
+    }
+
     return (
-        <DescriptionList isCompact isHorizontal>
-            <ContainerImage image={image} />
-            {(command?.length > 0 || args?.length > 0) && (
-                <>
-                    {command.length > 0 && (
-                        <DescriptionListItem
-                            term="Commands"
-                            desc={<MultilineDescription descArr={command} />}
-                            aria-label="Commands"
-                        />
-                    )}
-                    {args?.length > 0 && (
-                        <DescriptionListItem
-                            term="Arguments"
-                            desc={<MultilineDescription descArr={args} />}
-                            aria-label="Arguments"
-                        />
-                    )}
-                </>
-            )}
-            {!!resources && (
-                <DescriptionListItem
-                    term="Resources"
-                    desc={
-                        resources ? (
-                            <ContainerResourcesDescriptionList resources={resources} />
-                        ) : (
-                            'None'
-                        )
-                    }
-                />
-            )}
-            {!!volumes &&
-                (volumes.length === 0 ? (
-                    <DescriptionListItem term="volumes" desc="None" />
-                ) : (
-                    volumes.map((volume, i) => (
-                        <DescriptionListItem
-                            key={volume.name}
-                            term={`volumes[${i}]`}
-                            desc={<ContainerVolumeDescriptionList volume={volume} />}
-                        />
-                    ))
-                ))}
-            {!!secrets &&
-                (secrets.length === 0 ? (
-                    <DescriptionListItem term="secrets" desc="None" />
-                ) : (
-                    secrets.map((secret, i) => (
-                        <DescriptionListItem
-                            key={`${secret.name}_${secret.path}`}
-                            term={`secrets[${i}]`}
-                            desc={<ContainerSecretDescriptionList secret={secret} />}
-                        />
-                    ))
-                ))}
-        </DescriptionList>
+        <Card isFlat>
+            <CardTitle component="h3">Container configuration</CardTitle>
+            <CardBody>{content}</CardBody>
+        </Card>
     );
 }
